+++ TimeCtr.h
@@ -21,7 +21,7 @@ public:
 	bool actionSerial(char **, byte);
 	void actionTimer(unsigned long);
 	uint32_t getTimestamp();
-	uint32_t getTemp();
+	float getTemp();
 
 	static DS3231* clock;
